Refactor improve-digital into own file
